+++ src/lib/Features.ts
@@ -49,7 +49,6 @@ export const scrapUrl = async (
  */
 export const compareVersion = (
   name: string,
-  uuid: string,
   sourceCodeVersion: string,
   productionVersion: string,
   urlGitHub: string,
@@ -75,7 +74,6 @@ export const compareVersion = (
   }
   return {
     name: name,
-    uuid: uuid,
     githubLatestRelease: sourceCodeVersion,
     productionVersion: productionVersion,
     state: uptodateState,
@@ -140,7 +138,6 @@ export const getUpToDateOrNotState = async (
       resolv(
         compareVersion(
           record.name,
-          record.uuid,
           githubVersion || "",
           productionVersion || "",
           record.urlGitHub || "",
